Moved lexical GridSelection to the tables package
