+++ packages/lexical-playground/src/plugins/TableCellResizer/index.tsx
@@ -18,13 +18,13 @@ import {
   $getTableRowIndexFromTableCellNode,
   $isTableCellNode,
   $isTableRowNode,
+  DEPRECATED_$isGridSelection,
   getCellFromTarget,
 } from '@lexical/table';
 import {
   $getNearestNodeFromDOMNode,
   $getSelection,
   COMMAND_PRIORITY_HIGH,
-  DEPRECATED_$isGridSelection,
   SELECTION_CHANGE_COMMAND,
 } from 'lexical';
 import * as React from 'react';
