+++ packages/lexical/src/__tests__/unit/LexicalEditor.test.tsx
@@ -33,7 +33,6 @@ import {
   COMMAND_PRIORITY_EDITOR,
   COMMAND_PRIORITY_LOW,
   createCommand,
-  DEPRECATED_$createGridSelection,
   ElementNode,
   ParagraphNode,
   TextNode,
@@ -1158,90 +1157,6 @@ describe('LexicalEditor tests', () => {
         const stringifiedEditorState = JSON.stringify(
           editor.getEditorState().toJSON(),
         );
-        parsedEditorState = editor.parseEditorState(stringifiedEditorState);
-        parsedEditorState.read(() => {
-          parsedRoot = $getRoot();
-          parsedParagraph = parsedRoot.getFirstChild();
-          paragraphKey = parsedParagraph.getKey();
-          parsedText = parsedParagraph.getFirstChild();
-          textKey = parsedText.getKey();
-        });
-      });
-
-      it('Parses the nodes of a stringified editor state', async () => {
-        expect(parsedRoot).toEqual({
-          __cachedText: null,
-          __dir: 'ltr',
-          __first: paragraphKey,
-          __format: 0,
-          __indent: 0,
-          __key: 'root',
-          __last: paragraphKey,
-          __next: null,
-          __parent: null,
-          __prev: null,
-          __size: 1,
-          __type: 'root',
-        });
-        expect(parsedParagraph).toEqual({
-          __dir: 'ltr',
-          __first: textKey,
-          __format: 0,
-          __indent: 0,
-          __key: paragraphKey,
-          __last: textKey,
-          __next: null,
-          __parent: 'root',
-          __prev: null,
-          __size: 1,
-          __type: 'paragraph',
-        });
-        expect(parsedText).toEqual({
-          __detail: 0,
-          __format: 0,
-          __key: textKey,
-          __mode: 0,
-          __next: null,
-          __parent: paragraphKey,
-          __prev: null,
-          __style: '',
-          __text: 'Hello world',
-          __type: 'text',
-        });
-      });
-
-      it('Parses the text content of the editor state', async () => {
-        expect(parsedEditorState.read(() => $getRoot().__cachedText)).toBe(
-          null,
-        );
-        expect(parsedEditorState.read(() => $getRoot().getTextContent())).toBe(
-          'Hello world',
-        );
-      });
-    });
-
-    describe('grid selection', () => {
-      beforeEach(async () => {
-        init();
-
-        await update(() => {
-          const paragraph = $createParagraphNode();
-          originalText = $createTextNode('Hello world');
-          const selection = DEPRECATED_$createGridSelection();
-          selection.set(
-            originalText.getKey(),
-            originalText.getKey(),
-            originalText.getKey(),
-          );
-          $setSelection(selection);
-          paragraph.append(originalText);
-          $getRoot().append(paragraph);
-        });
-
-        const stringifiedEditorState = JSON.stringify(
-          editor.getEditorState().toJSON(),
-        );
-
         parsedEditorState = editor.parseEditorState(stringifiedEditorState);
         parsedEditorState.read(() => {
           parsedRoot = $getRoot();
